Integration tests mock
