+++ .env.test
@@ -1,12 +1,14 @@
+FLASK_SKIP_DOTENV=True
+
 # Flask server configuration.
 FLASK_APP=todo_app/app
 FLASK_ENV=development
 
 # Change the following values for local development.
+
 SECRET_KEY=secret-key
-TOKEN_KEY=Za.M)X6mM&-f`2!S
-API_KEY=![fUS9d)5,TW6=X
+TOKEN_KEY=Za471a84fdd4e15b6202edcfd93547d3c57311cee3f8a93a82b85aac1f6a7c1f
+API_KEY=9af06dec5e8bdd01f5b12fTW6fd93
 TODO_LISTID= 6047ecb966a18d23e8029cd2
 DOING_LISTID = 6047ecb966a18d23e8029cd3
 DONE_LISTID = 6047ecb966a18d23e8029cd4
-FLASK_SKIP_DOTENV=True
